linkmageddon!
Time to get rid of the old, salvage what's needed and move onto the new surface.

`links` is basically a mgmt link paired with either a sender or receiver link, and it handles doing the right thing for Recover(), Close(), etc.. (basically supercedes linkState)

There's a bunch of litle picky things to get rid of (a little message translation left!) but for the most part it should be okay.

The tests are completely devastated by these changes as they were mocking and stubbing a lot of the old structures. Haven't worked it completely out but I think `links` will be a good waystation for stubbed types (for the most part, nobody knows they're actually using AMQP).

I also moved the settlement logic into a little struct (`settler`) but it's bugging me a bit. Really I just want to use the high level Receiver in processor but I think I need to expose just a teeny bit of extra detail from Receiver to do that (right now, it's only exposing 'receiveBatch' and that would probably add more latency than I want just to receive a single message at a time.

See you on Monday, dearest repo.